Use an EnvironmentFile for extra variables.
This renders the common extra variables (including JUJU_UNIT and
SERVICE_ENVIRONMENT among others) defined by the OLS layer (and possibly
others) into a file in the service's etc/ directory. This file can
trivially be used via EnvironmentFile in the systemd unit (also done
here) or sourced manually if the service needs to be diagnosed or
otherwise run manually.

This commit doesn't move the existing variables, just renders the new
ones in the new file.